pyproject.toml
[tool.poetry]
name = "telegram_api_bot"
version = "0.1.0"
description = ""
authors = ["un1ck-andy <105160048+un1ck-andy@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyTelegramBotAPI = "^4.7.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
